requirements.txt
# Core dependencies
langchain>=0.1.8
langchain-community>=0.0.19
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langchain-chroma>=0.1.0
openai>=1.10.0
anthropic>=0.23.3
chroma-hnswlib>=0.7.3
chromadb>=0.4.22
faiss-cpu>=1.7.4
pydantic>=2.5.3
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pandera>=0.17.2
pandas>=2.1.3
typer>=0.9.0
tqdm>=4.66.1
rich>=13.6.0

# Less expensive models
llama-cpp-python>=0.2.32
hf-transfer>=0.1.4
transformers>=4.36.2
torch>=2.1.2

# Web application
fastapi>=0.105.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
markdown>=3.5.2
bleach>=6.1.0
libsass>=0.23.0

# Development dependencies
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
ruff>=0.1.5